Refactored the previous definition of a ChARGe experiment to a task. (#60)
* Refactored the previous definition of a ChARGe experiment to a task.
This better represents the current planning, where a task is a set of
steps, using a mixture of prompts and tools, to accomplish a goal.  A
(future) experiment class will be composed of one or more tasks and
include the state of the system to provide a complete representation
of the SMEs goals.

* Apply suggestions from code review

Co-authored-by: Shehtab Zaman <szaman5@binghamton.edu>

* Update README.md

Co-authored-by: Shehtab Zaman <szaman5@binghamton.edu>

* Fixed missed variables.

* Apply suggestions from code review

Co-authored-by: Shehtab Zaman <szaman5@binghamton.edu>

---------

Co-authored-by: Shehtab Zaman <szaman5@binghamton.edu>